requirements.txt
streamlit
pandas
numpy
plotly
yfinance
feedparser
Pillow
